feat(core): add --mock-browser-env flag to `sanity exec` command (#2319)
