+++ packages/@sanity/core/src/actions/exec/registerBrowserEnv.js
@@ -1,0 +1,6 @@
+const mockBrowserEnvironment = require('../../util/mockBrowserEnvironment')
+
+mockBrowserEnvironment(
+  // eslint-disable-next-line no-process-env
+  process.env.SANITY_BASE_PATH
+)
