+++ packages/@sanity/core/src/actions/graphql/getSanitySchema.js
@@ -1,75 +1,13 @@
-const pirates = require('pirates')
-const jsdomGlobal = require('jsdom-global')
-const pluginLoader = require('@sanity/plugin-loader')
-const requireContext = require('../../util/requireContext')
-const registerBabelLoader = require('./registerBabelLoader')
-
-const getFakeGlobals = () => ({
-  __DEV__: false,
-  requestAnimationFrame: (cb) => setTimeout(cb, 0),
-  cancelAnimationFrame: (timer) => clearTimeout(timer),
-  InputEvent: global.window && global.window.InputEvent,
-})
-
-function provideFakeGlobals() {
-  const fakeGlobals = getFakeGlobals()
-  const stubbedKeys = []
-  Object.keys(fakeGlobals).forEach((key) => {
-    if (!global[key]) {
-      global[key] = fakeGlobals[key]
-      stubbedKeys.push(key)
-    }
-  })
-
-  return () => {
-    stubbedKeys.forEach((key) => {
-      delete global[key]
-    })
-  }
-}
+const mockBrowserEnvironment = require('../../util/mockBrowserEnvironment')
 
 function getSanitySchema(basePath) {
-  const domCleanup = jsdomGlobal()
-  const windowCleanup = () => global.window.close()
-  const globalCleanup = provideFakeGlobals()
-  const contextCleanup = requireContext.register()
-  const cleanupFileLoader = pirates.addHook(
-    (code, filename) => `module.exports = ${JSON.stringify(filename)}`,
-    {
-      ignoreNodeModules: false,
-      exts: getFileExtensions(),
-    }
-  )
-
-  registerBabelLoader(basePath)
-  pluginLoader({basePath, stubCss: true})
+  const cleanup = mockBrowserEnvironment(basePath)
 
   const schemaMod = require('part:@sanity/base/schema')
   const schema = schemaMod.default || schemaMod
 
-  cleanupFileLoader()
-  contextCleanup()
-  globalCleanup()
-  windowCleanup()
-  domCleanup()
-
+  cleanup()
   return schema
-}
-
-function getFileExtensions() {
-  return [
-    '.jpeg',
-    '.jpg',
-    '.png',
-    '.gif',
-    '.svg',
-    '.webp',
-    '.woff',
-    '.woff2',
-    '.ttf',
-    '.eot',
-    '.otf',
-  ]
 }
 
 module.exports = getSanitySchema
