+++ packages/@sanity/core/src/util/registerBabelLoader.js
